Replaced all images with urls and refactored arrays
