+++ public/index.html
@@ -7,9 +7,10 @@
   <meta name="viewport" content="width=device-width, initial-scale=1" />
   <meta name="theme-color" content="#000000" />
   <meta property="og:title" content="Fret" />
+  <meta property="og:image"
+    content="https://user-images.githubusercontent.com/30041751/241193789-796ffdfc-8cfd-493b-a197-c325b4acd5ae.jpg" />
   <meta property="og:description"
     content="Fret is an instrumental trio from Nepal, creating new form of music, furnishing fusion grooves in an evolved manner blending traditional and ethnic music with an influence of eastern classical, western music and world music through Sarangi, Flute and Newari drums." />
-  <meta property="og:image" content="https://fretofficial.com/static/media/music_for_all.c3ee24c1.jpg" />
 
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.2/dist/css/bootstrap.min.css" rel="stylesheet"
     integrity="sha384-uWxY/CJNBR+1zjPWmfnSnVxwRheevXITnMqoEIeG1LJrdI0GlVs/9cVSyPYXdcSF" crossorigin="anonymous">
