+++ src/Links.js
@@ -1,1 +1,2 @@
+export const imageBaseUrl = "https://user-images.githubusercontent.com/30041751/";
 export const registrationLink = "https://docs.google.com/forms/d/e/1FAIpQLSetP5Qyk8TdHtUNLmxcUrg5KysfD7RMxtLciR03IeddLZDSBg/viewform?fbclid=IwAR04CXs7uXHrPfHpTLrb9gqsZmzB9I4_EXmsqkA5El9QmySm89V5OTi9IBw";
