+++ src/components/FeatureVideo.js
@@ -1,9 +1,8 @@
-import featured_video from "../videos/featured_video.mp4"
-
 const FeaturedVideo = () => {
+  const videoLink = "https://user-images.githubusercontent.com/30041751/241194586-599f793c-e44b-405c-b95b-7039e9f1b4c9.mp4";
   return (
     <video width="100%" controls autoplay>
-      <source src={featured_video} type="video/mp4" />
+      <source src={videoLink} type="video/mp4" />
       Sorry, your browser doesn't support videos.
     </video>
   );
