+++ src/components/initiations/InitiationsData.js
@@ -1,0 +1,33 @@
+const initiations = [
+  {
+    "id": 1,
+    "title": "Ambience of Bhaktapur",
+    "images": [
+      "241193923-0ac30206-1e56-4b23-99a6-912640cc6a44.png",
+      "241193937-294ad049-581f-4528-81b6-984ac21eb0ce.jpg",
+      "241193939-9e9cbe34-6de2-41db-8e60-63f5c7045155.jpg",
+      "241193942-f37ca15c-923b-4cf6-ae9b-858e3f0e2308.jpg"
+    ],
+    "details": <p>
+      Bhaktapur is well known all over the world for its culture, heritage and
+      traditions. Within the culture of Bhaktapur, Musical culture and tradition
+      is an important aspect. Music is a very essential factor which keeps the
+      ambience alive in Bhaktapur. WIthout the existing traditional music,
+      Bhaktapur will lose its soul. So we have named this initiation as ‘Ambience
+      of Bhaktapur’. Nowadays, we can see our young generation are getting
+      detached from their own origin, of which music is a dominant aspect. To keep
+      the ambience of Bhaktapur alive, music should be understood and learned. So,
+      in order to understand anything, we have to spread knowledge about it. So
+      the most effective way of doing so is to give it an academic direction.
+      ‘Ambience of Bhaktapur’ is an initiation designed by Fret with a motive to
+      make the music and musical instrument known among students from their school
+      level. We believe giving musical knowledge about music from school level
+      will develop an interest to learn music and even encourage them to pursue
+      music as a profession too. Through this initiative we firmly believe to
+      promote our music and musical instruments into the world music scenario.That
+      is why we are taking initiatives through schools of Bhaktapur.
+    </p>
+  }
+]
+
+export default initiations;
