+++ src/components/videos/VideosData.json
@@ -1,0 +1,14 @@
+[
+  {
+    "title": "Solitude",
+    "link": "https://www.youtube.com/embed/Xed9YL3tXiU"
+  },
+  {
+    "title": "Jiwanta",
+    "link": "https://www.youtube.com/embed/_cqPL7Lvi50"
+  },
+  {
+    "title": "Holi Instrumental",
+    "link": "https://www.youtube.com/embed/THPkX2oq-4c"
+  }
+]
